+++ Frontend/src/app/layout.tsx
@@ -2,6 +2,7 @@ import "./globals.css";
 import type { Metadata } from "next";
 import BackgroundContainer from "@/components/ui/BackgroundContainer";
 import Header from "@/components/ui/Header";
+import { ResultContextProvider } from "@/context/ResultContext";
 export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
@@ -11,9 +12,11 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html lang="en">
       <body>
-        <Header />
-        {children}
-        <div id="modal"></div>
+        <ResultContextProvider>
+          <Header />
+          {children}
+          <div id="modal"></div>
+        </ResultContextProvider>
       </body>
     </html>
   );
