+++ Frontend/src/components/AnswerRecord/AnswerRecord.module.css
@@ -53,23 +53,6 @@
   font-weight: 800;
 }
 
-.side_button {
-  width: 46px;
-  height: 92px;
-}
-
-.button_container {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  align-items: center;
-  letter-spacing: 0px;
-  color: #1e1e1e;
-  opacity: 1;
-  font-size: 20px;
-  font-weight: 700;
-}
-
 .dark {
   background: #c2c2c2 0% 0% no-repeat padding-box;
   opacity: 0.81;
