+++ Frontend/src/components/ui/Recorder/Recorder.module.css
@@ -1,0 +1,36 @@
+.content_container {
+  display: flex;
+  width: 100%;
+
+  justify-content: space-between;
+}
+
+.content {
+  padding-top: 120px;
+}
+
+.recorder_container {
+  margin-top: 38px;
+  display: flex;
+  align-items: center;
+  gap: 58px;
+}
+
+.recorder_button {
+  padding: 23px;
+  background: #f3f3f3 0% 0% no-repeat padding-box;
+  opacity: 1;
+  border-radius: 100%;
+}
+
+.button_container {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  align-items: center;
+  letter-spacing: 0px;
+  color: #1e1e1e;
+  opacity: 1;
+  font-size: 20px;
+  font-weight: 700;
+}
